# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation for Mahasiswa",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/mahasiswa": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Dapatkan daftar semua mahasiswa",
                "description": "Mengembalikan daftar semua mahasiswa",
                "operationId": "getMahasiswa",
                "responses": {
                    "200": {
                        "description": "Daftar mahasiswa berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mahasiswa"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ada data mahasiswa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data mahasiswa tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Tambah mahasiswa baru",
                "description": "Membuat data mahasiswa baru",
                "operationId": "createMahasiswa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "nim",
                                    "email",
                                    "jurusan"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "nim": {
                                        "type": "string",
                                        "example": "2024004"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@university.com"
                                    },
                                    "jurusan": {
                                        "type": "string",
                                        "example": "Teknik Informatika"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mahasiswa berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mahasiswa"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Input tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "properties": {
                                                "nama": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "nim": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "jurusan": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mahasiswa/{id}": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Dapatkan detail mahasiswa",
                "description": "Mengembalikan data mahasiswa berdasarkan ID",
                "operationId": "getMahasiswaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data mahasiswa ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mahasiswa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mahasiswa tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mahasiswa tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Update data mahasiswa",
                "description": "Mengupdate data mahasiswa berdasarkan ID",
                "operationId": "updateMahasiswa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "nim",
                                    "email",
                                    "jurusan"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "nim": {
                                        "type": "string",
                                        "example": "2024004"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@university.com"
                                    },
                                    "jurusan": {
                                        "type": "string",
                                        "example": "Teknik Informatika"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data mahasiswa berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mahasiswa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mahasiswa tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mahasiswa tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Input tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "properties": {
                                                "nama": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "nim": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "jurusan": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Hapus data mahasiswa",
                "description": "Menghapus data mahasiswa berdasarkan ID",
                "operationId": "deleteMahasiswa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data mahasiswa berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mahasiswa berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mahasiswa tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mahasiswa tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operasi berhasil"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "properties": {
                            "field": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Mahasiswa": {
                "title": "Mahasiswa",
                "description": "Model Mahasiswa",
                "properties": {
                    "id": {
                        "description": "ID mahasiswa",
                        "type": "integer"
                    },
                    "nama": {
                        "description": "Nama mahasiswa",
                        "type": "string"
                    },
                    "nim": {
                        "description": "Nomor Induk Mahasiswa",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email mahasiswa",
                        "type": "string"
                    },
                    "jurusan": {
                        "description": "Jurusan mahasiswa",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Mahasiswa",
            "description": "API Endpoints untuk manajemen data mahasiswa"
        }
    ]
}